physicalTwinDriver/requirements.txt
async-timeout==4.0.2
Deprecated==1.2.13
packaging==21.3
pyparsing==3.0.8
pyserial==3.5
redis==4.2.2
wrapt==1.14.0
